Move logging to shared/
