+++ src/api/index.js
@@ -39,11 +39,10 @@ import bleDriverJs from 'pc-ble-driver-js';
 import nrfjprog from 'pc-nrfjprog-js';
 import usb from 'usb';
 import electron from 'electron';
-import { logger } from './logging';
+import { logger } from '../shared';
 import * as core from './core';
 
 const bleDriver = bleDriverJs.api ? bleDriverJs.api : bleDriverJs;
-core.logger = logger;
 
 export {
     bleDriver,
