+++ src/shared/logging/__tests__/appTransport-test.js
