+++ src/shared/logging/appTransport.js
