+++ src/shared/logging/consoleTransport.js
