+++ src/shared/logging/index.js
